requirements.txt
yfinance
pandas
plotly
python-dateutil
pytz
